moved sql and mongo scripts to their own folders
