--- models/index.js
@@ -1,5 +1,0 @@
-module.exports = {
-    Professional: require('./professional'),
-    Student: require('./student'),
-    School: require('./school')
-}
