--- models/sequelize/student.js
@@ -1,1 +1,0 @@
-//TODO: Create a Sequelize model for Students
