--- models/student.js
@@ -1,11 +1,0 @@
-const mongoose = require('mongoose');
-const Schema = mongoose.Schema;
-const studentSchema = new Schema({
-    firstName:{ type: String, required: true },
-    lastName:{ type: String, required: true },
-    email: {type: String, required: true},
-    // joinDate: {type: Date, default: Date.now},
-});
-
-const Student = mongoose.model("Student", studentSchema);
-module.exports = Student;
