--- routes/api/config/config.json
@@ -1,21 +1,0 @@
-{
-    "development": {
-        "username": "root",
-        "password": "root",
-        "database": "pc_builder",
-        "host": "localhost",
-        "dialect": "mysql"
-    },
-    "test": {
-        "username": "root",
-        "password": null,
-        "database": "builds_test",
-        "host": "localhost",
-        "dialect": "mysql",
-        "logging": false
-    },
-    "production": {
-        "use_env_value": "JAWSDB_URL",
-        "dialect": "mysql"
-    }
-}
